[po-82] Allow campaign creation not linked to any organization
This is a first step to the campaign creation.
We allow first to create a campaign not linked to any organizations.
Linking the campaign to an organization needs some extra code that will make this user story heavier and longer.
In order to avoid a tunnel effect (we are already in), we decided to push the functionality as is, and to make the link between the campaign and the organization in another story coming just after this one :)