CirleCI: always use a per-commit cache
The shared node_modules caches were partially disabled in 23dffdc,
this completes the work by making sure every cache is keyed using
the current branch and SHA1.

It would probably be better to either use the CircleCI workspace
for this per-workflow data, or go back to shared caches but making
sure they are safe to use, i.e. keying on package-lock.json in addition
to package.json.