[pf-359] [BSR] CircleCI: always use a per-workflow cache and simplify workflow
The shared node_modules caches were partially disabled in 23dffdc,
this completes the work by making sure every cache is keyed using
the current workflow id.

It would probably be better to either use the CircleCI workspace
for this per-workflow data, or go back to shared caches but making
sure they are safe to use, i.e. keying on package-lock.json in addition
to package.json.

Also, because `npm install` at the root triggers `npm install` for all
subdirectories, it is redundant to run `npm install` again in each
one, so the `install_` jobs were removed.

It would be more efficient to run the `npm install`s in parallel jobs,
but since we can't prevent the root `npm install` from triggering the
subdirectories (the Scalingo deployment currently depends on this behavior),
we have to settle for not doing it twice.