CircleCI: install all dependencies only once
Because `npm install` at the root triggers `npm install` for all
subdirectories, it is redundant to run `npm install` again in each
one.

It would be more efficient to run the `npm install`s in parallel jobs,
but since we can't prevent the root `npm install` from triggering the
subdirectories (the Scalingo deployment currently depends on this behavior),
we have to settle for not doing it twice.