[pf-362] api: generate SAML requests and consume SAML responses
/api/saml/login redirects the user to the IDP with a SAML request in the
URL.

The IDP processes the request and sets up a POST from the user's browser
to /api/saml/assert (assuming this endpoint was specified in the SP
metadata we served to the IDP at /api/saml/metadata.xml).

/api/saml/assert looks for an existing user based on the "mail"
attribute, or the "IDO" attribute that is provided by the GAR,
searching for a match on the "samlId" attribute in our database.

If no such user exists one is created using the first and last name
attributes found in the SAML response, saving of course the "samlId".

A login token is then generated for the user and passed to the front
application in the URL.