[pf-416] add locking strategy to redis cache get
Now, when we try to get something from the cache but it is not available,
the cache is going to lock the key, retrieve the value using the
generator function given to cache.get, and add the value into the cache.

Redis redlock mecanism is used to lock the cache key.