[pf-362] Insert a null email in database for SAML users
Fixes not being able to create more than one SAML-based user account.

Previously, the code was inserting an empty string in the
email column for users automatically created via SAML, which
tripped the unicity constraint on the email column after the
first user.

We now make sure those users are created with a null email,
which requires making the email column nullable.