[pf-143] Redirect to assessments.resume when creating or resuming a course assessment
This fixes resuming a certification when it is completed (PF-455), and
removes a bit of duplication of the "next challenge" logic.

The legacy certifications.resume route (/certifications/1234) now goes
directly to courses.create-assessment

Since assessments.resume is now used for new assessments as well, the
loading message was changed to be more generic.

Also, if an error occurred while retrieving an existing assessment, it
would not be shown because the redirect hook is expected to be
synchronous. Using the afterModel hook instead for the same purpose
makes Ember wait for promise resolution before showing the page.