Add a test method to clear all data from tables
🌍 Context
==========

We use the `DatabaseBuilder` to generate and clean data for tests, but
with the limitation that it could only clean the data that it had
created, it could not clean data created by other means.

🔧 Fix
======

This commit adds a `clearAllTables` method to the DatabaseBuilder that
clears all data from all tables, except for tooling & system tables
(`knex_*` tables for instance).