Change /api/scripts to avoid side effects in tests
🌍 Context
==========

Many scripts in `/api/scripts` are required by tests (because they are tested), and had side effects if `NODE_ENV` was != `'test'`. This was a problem to

🔧 Fix
======

There exist in Node.js a better protection mecanism to avoid side effects when standalone scripts are required instead of run.

So in `/api/scripts`, I changed all instances of:

```
if (process.env.NODE_ENV !== 'test') {
  main();
}
```

to:

```
if (require.main === module) {
  main();
}
```