factoring all resume process in assessments.resume route
The route assessments.resume handle to answer the question "what is the next page to display?" based on whether or not there is still challenges to ask, whether or not assessment is completed, whether or not there is checkpoints to display.
This route handle it for all kind of assessments : CERTIFICATION, PLACEMENT, DEMO, SMART_PLACEMENT.

DEMO were the only assessments that uses challengeId when querying next challenge, but it is not required since we can search for already answered questions. The way to search for next challenge is thus changed.