Optimize Circle CI build
General cleanup of the Circle CI config.
We now use the cache more aggressively to reuse downloaded npm packages
between builds.

`npm install` at the root will no longer automatically trigger
installation in subdirectories. This can still be triggered
by running `npm run ci:all` at the root.

Also, we now use Circle CI's built-in parallelism feature to
split the mon-pix job in three instances. The build/test split
for this app was undone: now the app is rebuilt in every
container. This does mean extra work but given the cost of
transporting the built files between containers, it's a wash
in terms of latency and the configuration is simpler.