[TECH] Add var to start API in Cypress, fix a test
The `AUTH_SECRET` env var was missing in the API start env, so the token
generation was failing (=> impossible to login).

Note: other missing env vars prevent the API to function correctly (Airtable
access). Let's see how to inject secrets in CircleCI (or switch to
Travis?)

And a test was failing due to wrong CSS selector for submit button (how
to write better tests?)