[TECH] Make Unit tests pass without the database
Some Unit tests where using the database, either because they are
integration tests or because some function where not stubed.

This does not allow to run unit tests in isolation of the database and
does not correspond to the definition of a unit test.

This commit moves integration tests to the integration dir, or stubs the
method that make a real unit tests.