Specifiy caching behaviour for nginx servers
Currently, requests that return the contents of index.html do not
specify a caching policy, with the result that some clients cache
the index.html contents and can be stuck with an old version of the
front applications.

This change directs browsers never to cache responses that serve
index.html. Files in /assets/ are safe to cache forever on the other
hand, since they have unique names.