[po-196] Added an options parameter to campaign participation serializer to work seamlessly with included relationshps or links
If the serializer computes a relationship link but does not have any id to that relationship,
such as with dynamic relationships that are recomputed on the fly everytime, then EmberData will
not follow the link. Therefore, we need to be able to toggle the 'ignoreRelationshipData' boolean.
When set to false, EmberData will see the included relationships and wont fetch them one by one.
When set to true, EmberData will follow the sole link to the relationship and not ignore it.