Use _.isNil method in database builder factories
A common pattern in javascript to provide default values is to write
val = val || defaultVal. This works great, except when the val is the
value 0, in which case js considers it falsy and provided the default value.

The fix is to use the _.isNil lodash method to only provide a default value
when val is null or undefined;