[pf-579] Move initialization of competence in the model
When fetching the profile, the competences were
augented with default values for 'level' and 'status'
since #581 ['add a second chance button'].

This initialization is only useful for the status key, used
by the frontend to display the correct 'Start' button.

Moving this initialization is the constructor will guarantee
that there are always default values, which is something we
want.