[po-199] Add new analytics domain concepts.
The two concepts added in api/lib/domain/models are CampaignCollectiveResult and CampaignCompetenceCollectiveResult.

A CampaignCollectiveResult is a container of CampaignCompetenceCollectiveResult objects. It has an ID that is the same as the campaign it refers to (for store-based front applications).

A CampaignCompetenceCollectiveResult is an object that describe the collective result (total number of targeted skills covered by the campaign, average of validated skills by participants who shared their contribution, etc.). It also has an ID as followed : <campaignId>_<competenceId>.