[po-199] Add CampaignCollectiveResultRepository.getCampaignCollectiveResult().
This module/method concentrates the major business logic of the feature.

The algorithm follows 4 steps :

1. Fetch data : we get the data from the data sources (Airtable, SQL DB), i.e. the campaign, its target profile with targeted skills, its associated particpiations, the participants, their knowldege elements and the standards competences & skills
2. Filter data : we keep only the shared participations and the particpants positive knowledge elements concerned by the campaign targeted (profile) skills (+ some other filtering rules)
3. Forge data : we merge the data, based on the competence IDs
4. Format data : we build and return a CampaignCollectiveResult object with its array of CampaignCompetenceCollectiveResults

In order to maximise performances and ease of use, we decided to act closer to the data source. It's why use a lot of AirtableRecord objects (via Airtable data sources) and Bookshelf Model objects (vive la fonctionnalité "related" !).

If we need it in the future, we can make an SQL raw query to fetch Knowledge Elements instead of using related queries.

In order to make the code clearer/easier to shape (& test), we put some filtering logic directly in the Bookshelf Knowledge Element model object.