Update NPM version in all projects.
According to Node 10.15.3 (LTS) we are currently in NPM v6.4.1. Few days
ago, NPM released the 6.9.0 (cf. https://npm.community/t/release-npm-6-9-0/5911).

Upgrading the NPM version will add a step in the environment
installation. But not doing so, display a waring during each `npm
install` command execution.

In the next steps, we will try to upgrade Node version (with the risk of
leaving a comfortable LTS version). If the upgrade does not work, we
will have to take a decision about the version we keep of Node / NPM.

Maybe the PR associated to this commit will never finished in production. 🤷‍♀️