[pf-579] Refactor data-fetcher to use spread
The spread operator has interesting use cases when some code
moves a lot of dependencies around. It enhances readability
by reducing the amount of noise. However, the benefit does not
always outweight the cost which is loosing sight of which objects
are being passed around.

By documenting the signature with destructured
parameters and using the spread operator on the argument keywords, we
arguably get the best of both worlds.