[pf-579] Extract a private method in data-fetcher
The usage of Promise.all to parallelize async tasks forced
the code to use a weird hack in order to do it all in one
batch. Extracting a private method enhances readability by removing
the second Promise.all() call that was only here to ensure we don't
loose track of the target profile's skills down the promise chain.