api: fix start:watch script work with node-heapdump
When node-heapdump is `require()`'d it sets up a signal handler for
`SIGUSR2` that generates a heap dump. But `nodemon` also sends a
`SIGUSR2` signal to restart the application when it detects a change in
source files.

This changes the signal used by `nodemon` to `SIGTERM` to avoid this
conflict.