[pf-579] Add a migration to add competenceId on assessments
The previous design where competence evaluations referenced assessments
but not the other way around made it impossible to connect past
assessments to their competence evaluations after each reset.

The solution is not to add a competenceEvaluationId on assessment but
directly the competenceId. By doing so, we also improve the simplicity
of the get-next-challenge logic.

From a relationnal's point of view, since each competence based assessments
have a competence, these assessments will now have a competenceId
attribute.