Fix multiple opening of certification session report.
Before this commit, it was not possible to open several times the certification session report.

The root cause of the bug was the misuse of the input file uploader (and its handler).

To solve the problem, we replaced the previous mechanism with the preconised one based on addon component {{file-upload}}.