Fix a bug where the last candidate was ignored
readAttendanceSheet is refactored to get all the candidates with a lastName, instead of trying to find the last line of the file.
To do that, we extracted the code which parse the ODS file in another service.
The parsing method returns a json with all the lines below the header of the file, then readAttendanceSheet filters and formats the remaining candidates