[pf-788] Move signalling to build task instead of scalingo-post-ra-creation
The scalingo-post-ra-creation hook executes in the context of the deployed app.

We don't want to require the deployed app to have the node_modules/ directory
that is only useful during the build.

It's OK to run the signalling scripts repeatedly, since they do nothing if
a comment has already been posted.