[pc-67] Unplug the possibility to certify V1 profiles
Because of the V1->V2 migration, it should not be possible for users to
certify their V1 profiles anymore. However, if the V1 profile score was
superior to the V2 profile score, the code would still allow V1 profiles
to be certified. This commit will prevent this situation from happening
in the future.