[pf-773] Ceil pix score and level for certification
The result of a certification is stored in the competence-marks table, in
an 'append-only' manner. We store there all the information about the
current state of the certification algorithm, challenge, referential and
so on...

This way we ensure the certification results will never be
impacted. If we ever decide that we should also be able to recompute
dynamically past certifications, we will need a much more robust
versionning system.