[pf-881] Spread on an unnamed composite object
The current object is a big composite objects with data comming from
multiple sources, and some keys being renamed. The currrent structure is
questionnable and beyond the scope of the currrent refactoring as it
would imply rework on pix-admin too and is prone to regressions.

This commit seems like a good middleground. However, merging objects
like this take probably more properties than expected by the client. We
then rely on the Serializer to make sure we don't pass too many props,
which could on rare occasions be a problem. When we have a real domain
model for this object, this won't happen since constructor will filter
the keys it needs.