api: normalize structure of usecase modules
A few of the usecase modules did not follow the rule that the
canonical name for the exported function should appear as the
name of the function in the implementation file.

E.g. get-current-user.js should contain `function getCurrentUser(…)`.

This rule is meant to help when looking for the implementation
of a usecase using a text search (e.g. `grep`).

The list of import statements in `usecases/index.js` was regenerated
from the usecase modules using the following magical one-liner:

    (cd lib/domain/usecases &&
      content=$(< index.js awk '{print} /module.exports/ { exit }');
      (echo "$content";
       rg -r '$1' '.*exports.*function *([^(]+).*' *-*.js |
        rg -r "  \$2: require('./\$1')," '(.*)\.js:(.*)' |
        sort;
       echo "});") > index.js)