Isolate tests into different contexts
A test, that has a fundamentally different context from the others,
was nested into the main context. Unfortunately, the consequence
was that beforeEach hooks from both main context and specific
context were triggered, and both were sending a POST request
to the server. The promise of one of them was not waited on.