api: avoid Redis crash caused by backed-up queries
The node-redis client will by default maintain a backlog of Redis
queries whenever the Redis server is disconnected. We have observed
in production that in pathological cases, a temporary unavailability of
the Redis server could cause so many queries to be executed when Redis
went back up that it would immediately crash due to consuming all
available memory for client output buffers.