Improve User domain conversion in UserRepository.
Ember Data does not like anymore bidrectional relationships that are not
consistent.

If a User has memberships lazy loaded (thanks to a second
request to the API), then the memberships returned via
/api/users/:id/memberships MUST have a user ID well set.

If not, Ember Data will make its best to guess and fix the relationship
in its internal store, but it will before log a big BIG warning message in the
console.

In this commit, in the UserRepository, during the domain conversion
phase, after the User was instanced (with its relationships), we loop on
its memberships to set them the user.