Fix reset inconsistency
It appeared that when a competence was reset, no level was
displayed after a campaign participation. Knowledge-elements
were present though so we were able to display progression in
the circle.

Investigations revealed that the scorecard status should base
its status on knowledge-elements when there is no
competenceEvaluation or when the status is reset.