Make sure tables are empty when launching tests
This commit moves a target profile and its skills from a migration
file (where they don't belong) to a seed file.
It also uncomment an emptying database test that used to be
commented.
To finish, a before hook has been added to make sure all tables
are empty when launching tests.