Wrap the start certification in stepper component
This commit introduces a stepper component that leverages the
progression tracker, and verify that it works seemlessly with only one
step, the one that was already present (the certification starter).

In order to work, the stepper makes uses of dynamic "wrapper components"
that must expose specific keys, such as a component named main etc. More
features will come.