[pf-866] Start separate course from certifCourse
CertificationCourse is in debt. It starts with how it is
represented in an Assessment model API side.
This commit is a tiny first step toward the substantive work
on CertificationCourse. In the serialized version of Assessment,
the distinction between a course (demo, placement, etc) and
a certificationCourse has been added. Since sometimes the
courseId in assessment stands for a real certification course ID
record, we added a link to directly fetch the real data to a
valid endpoint. Lot of work needs to be done, we'll get there.