[pf-866] Add certificationCourse model App side
In a context of a certification test, assessment's courseId
is not just a random courseId, it refers to a real certification
course record in the database. We introduce here a distinction
between a course and a certificationCourse, and its related
data is now fetched thanks to a link provided by the JSON
API serialized Assessment.