Optimize data loading
To look up a challenge ID in the list of answers, we need to
have all the answers loaded, which can generate a lot of
unnecessary traffic.

This commit uses the `.hasMany(…).ids()` idiom of Ember Data
to retrieve the list of answer IDs from the assessment without
having to do an extra fetch.

Since we are already fetching the answer ID for the current
challenge, we then just have to look for that answer ID
within the list of answer IDs.

This results in being able to compute the current answer number
without any extra fetch.