Handle server errors explicitly in certification
The code currently transition away to the index route when the server
responds a 5xx error. This is problematic since such an error is
critical and should be reported back to the person in charge, during a
certification session. This commit displays an error message when such
an exceptionnal situation arises.