Make sure node is the main process in the container
When npm is used to start the process, it stays as the root
process in the container.

When the Scalingo infrastructure signals the container to stop,
it only signals the root process. But npm does not forward the
signal to its children, see https://github.com/npm/npm/issues/4603 .

To work around this we unfortunately cannot rely on an npm script,
we have to duplicate the definition of the "start" script into
the Procfile.