Add a PUT finalization endpoint for sessions
It has been decided to use a PUT for finalization the session for
various reasons. First, the PUT semantics is to be idempotent at the
protocol, why is exactly how we want this endpoint to be. Second, PUT is
used for updates and this is exactly what is is. Finally, the
sub-resource /finalizations allows for a payload-free endpoint, whereas
the response will be, the new updated resource.