Handles specific birthdate 'DD/MM/YYYY format
Certif team raised an alert that many PixCertif users
where copy pasting existing values into the attendance sheet
in order to register candidates into a session.
It seems like the file they are copying from presents
birthdate value with an extra ' at the beginning of it :
'25/12/2008   for example
We need to handle this properly as it seems to be an usual
practice.