Replace _.findLast with _.find
I strongly suspect the findLast call was used in the first place when
the repository did not filter knowledge elements that validate the same
skill with different date. Now, it does, thereby making this _.findLast
call unnecessary and, consequently, suspicious.

It also used the lodash shorthand syntax for better style.