Turn createOneLineOfCSV synchronous
The goal is to mutualize the fetching of all the necessary data in one
place to have only one asynchronous function and better see what's going
on.

The fetching of the user was also renamed to accomodate both the
retrieval and the check in order to avoid to fetch the user twice.