Add support for dynamic values over collections
The header property map must be able to have dynamically computed values
over collections of objects, where each object can be responsible for
multiple headers, hene the use of a _.flatMap.

When doing so, the information needed to dynamically compute these new
headers is often tightly linked to the value that will be present.
Therefore, we add here the possibility of having a 'value' instead of a
'propertyName' so we can use it directly in the csv-service.