Introduce a CampaignIndividualResult domain object
It is about time that all the aggregation done here to generate a csv of
the individual results for a campaign live on the same domain model
object. The name CampaignIndividualResult will find its way right next
to its plural-counterpart, the CampaignCollectiveResult, for
consistency.

This commit only adds the global part, taking into account only the data
that is common to every one. Therefore, it is not (yet) really
individual.