Upgrade Ember from 3.9 to 3.10 for Pix App
This commit suppresses the following elements :
- eslint no jQuery rule (temporary)
- testem disable GPU until good ember version ?

The following problems hava been encountered :
- mustache inputs are no more supported properly
and have been replaced by new angle bracket inputs
- href proprety in link to component are no more
generated if there is no router service in the
integration test. Also, an expect on the href
after a LinkTo is quite uncommon across the Pix
ecosystem and we thought that it resulted to
test EmberJs itself. Thus we deleted the expected.