Delete dirty tables instead of records
In databaseBuilder.clean method, we now completely delete
a table whenever it has been dirty. This avoid going through
each table to delete specifically an added record. We have
to work on clear databases on each test anyway.
For that we have to make sure that we run the DELETE sql command
accordingly to the order of dependency ruled by the fk constraints