Set property  to false in in-memory-cache.js in order to avoid clones and improve performances.
We mainly need to make read accesses, without mutate the stores values. So we do not need to clone objects.

According to the doc of node-cache : "(useClones) false is recommended if you want to achieve performance or save mutable objects or other complex types with mutability involved and wanted, because it'll only store references of your data." ~ https://www.npmjs.com/package/node-cache