Remove Redis client option enable_offline_queue
The "offline queue" feature was disabled when we put too much load on
Redis server.

It seems to be a problem when the API instance starts and listen to
requests meanwhile the Redis client connection is not yet established.

Since the improvement of Redis use (PR #922), we can enable offline queue again.