Render error when joining archived campaign
When the user attempts to join an archived campaign, the desired
behavior is to render a component explaining that the campaign was
archived. This will normally only be accessible when the user has
bookmarked the campaign link, or is replaying an older assessment for
which the browser's tab was still open.