Introduce findLatestOngoingByUserId method
The goal of this commit is to deal differently with the banner that is
displayed when an ongoing campaign has not been finished.

The profile page currently asks for all the campaign participations in
order to calculate on the frontend a lot of things, to know if there is
a campaign to continue. All of this computation should not be done here,
but rather be incapsulated in the API. This means that the getByUserId
repo method must completely change to filter out archived campaigns and
only keep the latest participation.

It also means to remove all the serialization logic that is currently
handling both the campaign retrieval, and the fact that this endpoint
can retrieve multiple campaign-participations. The url will also be
renamed to a more appropriate name to reflect this deep change.