Set validators with a synchronous behaviour
The 4 validators used in the API were not behaving the same.
Some where synchronous in success case, but async when in error,
and the other way around for the rest.
That explains the 500 type errors we have here and there API side.

We harmonize this by reinforcing the two following rules :
- All validators have a SYNCHRONOUS validate() function
- They either throw an error when in error,
    or return nothing (undefined) in success