Stop creating assessmentResult when not in certif
Regardless of the type of the assessment, we create an
assessmentResult anyway. For certification, this is mandatory
and at the core of the certification system. However, for the other
types of assessment, this is meaningless. It was only
generating an "empty" assessment result with no score whatsoever.
Let's stop creating dummy assessmentResult for anything besides
CERTIFICATION !