Clarify the ordering rule in database
The only rule should be the time, since two campaigns can never be
created at the same time, and since we don't want to distinguish between
multiple campaigns created the same day. Therefore, we verify that the
order is right, and rule out the possibility of a random database effect
by inserting another campaign, 'A', in the middle.