Fix hasSeenEndTestScreen value
HasSeenEndTestScreen value came from paper report (PV)
or from the finalization page in certif app.
When we import the PV we can put the value of hasSeenEndTestScreen in
the same attribute instead of create hasSeenEndTestScreenFromPaperReport.
So the downloadJuryFile can search the hasSeenEndTestScreen info in the same
way for paper report and from finalization page.