orga: bump fsevents sub-dependency for Node v12 support
`ember-auto-import` depends on `webpack` which depends on `watchpack`
which depends on `chokidar` which depends on `fsevents`.

The version of `fsevents` which was locked is 1.2.7 which does not
support Node v12. This results in voluminous error messages when
running `npm install` or `npm ci` in `orga` (with no further
consequence since `fsevents` is an optional dependency).

This commit upgrades `fsevents` to a more recent version which
supports Node V12. Since `fsevents` is not a direct dependency
of `orga`, it could not be updated with `npm update` so the
workaround used was to remove the `fsevents` block from
`package-lock.json` and run `npm install` again so that
a more recent version was picked and locked.