Add a findPaginatedSessions repository method
The current find is being replaced with a paginated-version heavily
inspired of the find organization of the same kind. However, the tests
are much more simplified by factoring out what is the common block for
each test set, thereby allowing each it() to only express in just a few
lines its main intent.

This commit also adds a candidatesCount model to the Session model and
let the repository populates it.