Get rid of useless constructor calls
First of all, calling constructor() is not recommended when
extending an Ember* class. Call to init() is preferred.
Secondly, a call to init() is necessary when initializing
a property that depends on other, or that requires an async call
for instance. It's not the case here, we just need to initialize
to a fixed value our properties