Replace challenges with certificationChallenges
Here again, let's not bring more confusion with business lexicon.
A challenge is what the content team is producing, meaning an object
with a question, possible answers, images etc....
A certificationChallenge wraps the afore mentionned item and exists
only in a certification context. It represent one of the questions
that will be asked to a specific candidate.
So, in certificationCourse, we indeed refer to the second one.