Fix filter search debounce on paginated list
Since PixOrga has been upgraded to Ember Octane,
the debounce was not working anymore.
This is because the controller has been turned into
an native JS Class instead of an Ember Object.
Thus, there is no need to invoke controller.set('attr', value)
to update a controller property. Directly doing
controller.attr = value now triggers everything correctly.
Since the input used to implement a filter search on
campaign name was a <Input> built-in Ember component,
the two way binding of this component to the
controller property triggers the new XHR request immediately
after hitting a key on the search field.
There are many solutions to this, the one picked here was
to switch for a classic <input> field, that is one-way binded