Fix buggy display when deleting filter name field
The computed that determines whether or not the
no-campaign-display should be displayed was depending
on two properties : 'hasNoCampaign' (that itself depends
on 'model') and 'name'.
Let's consider we were in a state where there was a name typed
in the search filter that returned 0 results.
Thing is, when deleting the search filter name, the computed
is being called twice really fast :
- the first time because name changed. At that point, name is empty,
and lets recall that model was ALSO empty, so the computed
displayNoCampaignPanel returns true (for a fraction of second)
- the second time because model changed. Indeed, since there are no
name filtered anymore, the API returned all the campaigns. So model is
not empty anymore, and displayNoCampaignPanel returns false (as it
should)
The solution is to change the computed dependency to observe solely
'model'. The changes on 'name' through the search filter DO trigger
a potential change on models.