[API] Add script forcing new termsOfService validation
This script update will update 75% of user's table, not far from
a million records. This may cause lock on table and queuing, promoting a
batch update solution.

We performed a dry-run on production-replica database from a week ago,
total time being less than 30 seconds. Support team agreed for such a
service interruption, so we keep this solution as-is.