Revert "Refactor read title from variable"
This reverts commit cf78e4a8ed192d74d7b3a5df0d1be3673e221fc5.

It turns out localizing page titles is more easily done in the
template rather than in a controller.

Controllers that existed only to provide a pageTitle property
are removed. `controllers/certifications/start.js` and
`controllers/profile.js` also provided `currentUser` but
it was no longer used in their templates so the controllers
were removed as well.