Revert "Refactor read title from variable"
This reverts commit 4fbe23a8b7a205efc76a1b8a4fd1477719591d05.

It turns out localizing page titles is more easily done in the
template rather than in a controller.

Controllers that existed only to provide a pageTitle property
are removed. `controllers/certifications/start.js` and
`controllers/profile.js` also provided `currentUser` but
it was no longer used in their templates so the controllers
were removed as well.