Remove pageTitle controller properties where unneeded
This is the same as the previous commit (move from pageTitle
controller properties back to {{title}} arguments in the template),
but for titles that were added after the pageTitle properties
were first introduced.

Only three non-trivial definitions of `pageTitle` remain.

Three more controllers can be deleted.

Note that the {{title}} element was entirely missing in
`terms-of-service.hbs`.