[API] Change base http error status to 400
When domain error are not wrapped by an http error, the API return an internal server error. Also, the stacktrace is not intercept by Sentry.
To fix that, we change the defaut status code to 400, we see that all domain error return an exception with status code 4XX.
To avoid the issue, Wrap all domain error by a defined http error.