Lint fix - Remove assignment use in mixin
error Use set(this, 'propertyName', 'value') instead of assignment for untracked properties that are used as computed property dependencies (or convert to using tracked properties) ember/no-assignment-of-untracked-properties-used-in-tracking-contexts

We chose to keep the use of the set method because it is in a mixin.
Mixin are supposed to dispear in time with glimmer; they are tolerated for now.