Force certif and deliveredAt date to Europe/Paris
The date on which the certification test was taken should be displayed using the timezone of the certification center in which it was held. As of today we do not have the timezone of the certification centers, the vast mojority of them being in France we choose to display those date in Europe/Paris time for the time being (pun intended).