refacto: change tooltip display condition
Before this commit, when the tooltip is displayed, it takes a css class 'hexagon-score__information--visible'. When it is hidden, the css class is changed to 'hexagon-score__information--hidden'. Instead if a class change, this commit will display the tooltip on a variable change: if 'displayHelp' is true the tooltip is displayed, if 'displayHelp' is false the tooltip is hidden.