feat(scss): remove negative margin to extract form from feedback panel
We noticed that rounded-panel import comes after challenge-item in app.scss.
It means that challenge-item style will always be overridden by rounded-panel.

As a consequence, we cannot use challenge-item as a specifier of rounded-panel and we
cannot apply BEM convention correctly. We could move components/rounded-panel to
globals/rounded-panel in order to always have rounded-panel on top of declarations.
However, such a redesign could break silently some visuals.

Hence, we added a modifier to rounded-panel instead of adding another
class to challenge-item.