Create migration script that fills in the bigint column from id for small databases
This migration is only executed for small databases (local, dev, integration, recette). It is why we make a test on MAX_ROW_COUNT_FOR_SYNCHRONOUS_MIGRATION.

We select 10 millions rows value because in production there are 700M rows.

Unfortunately, we did not find a pragmatic way to make tests on this migration.