[App] Add support to Matomo Session Recordings
In order to use Matomo plugin "Session Recordings", we need a way to keep old app versions static assets such as vendor.css and mon-pix.css. If not, the styles will not be displayed in Matomo, thus the session will not be exploitable.

A simple way to (nearly) fix the problem is to copy the generated assets during Ember build process as xxx.uncached.css ones and to define `data-matomo-href` attribute.

Links:
- https://matomo.org/faq/heatmap-session-recording/faq_24272/