fix(api): display badge's criteria in pix-admin
A badge criterion can be linked to skillIds that are 'périmés'.
However, when searching for the whole skill given the skillId, only the operative ('actif', 'archivé') skills are considered.
Thus, for skillId that are 'périmés' we cannot access its data, causing the GET badge request to fail.

In this commit, we prevent the failure by not considering the skillId of the badge criterion that are not operative.