fix(mon-pix): remove model based tracked property
hasFocusedOutWindow is a tracked property that corresponds to this.model.assessment.hasFocusedOutChallenge.
When the user leaves the window, this property is manually set to false, leading to a non-use of the model mentionned beforehand.
hasFocusedOutOfWindow can not be set back to its original state through the resetController method as it depends on the model.
We thus create a getter depending on the model that gets updated thanks to Ember reactivity system.

Co-authored-by: Alexandre Coin <alexandre.coin@pix.fr>
Co-authored-by: Matthias Ferraina <matthias